Fix token not refreshed when page refreshed
The logic behind this fix is quite simple.
I was able to fix it by calculating epoch time of the token expiry and store it in local storage.
Then, every time the page refreshed, the page will try to check the time difference between the current time and the expiry time.
The difference is used as a timeout delay. Once the timeout expires, the token will be refreshed.