Add seek and volume slider
This part is a little bit tricky to implement.
Volume slider is pretty much straightforward. Set the volume based on an input slider.
On the other hand, the seek and progress bar took a little bit longer to implement.
To achieve the progress bar animation, I need to use an interval which tracks the current position and increment it by 1 each second.
My first attempt was to set the interval everytime the user toggle the pause/play button. This way, the interval will only started when the user start the music.
But, this implementation was really flawed because the interval will never triggered if the user tries to seek the music.
To fix this, I figured that I need to check and listen to the state directly, rather than the button that triggers the pause.
I moved the implementation so that it will be triggered every time the state changes. Unfortunately, this creates duplicated interval.
This happens because the state event listener is on an useEffect hook. Then, I tried to store the interval id in a ref so that I can remove the interval whenever the music is paused and the interval is already instatiated.
This implementation still not worked because the state changes randomly (not only when the music is paused, the sdk does it randomly). Sometimes it caused the interval to be cleared because the last state trigger the remove interval function.
Eventually, I managed to implement the seek and progress bar by wrapping the implementation in a separate useEffect hook which fires everytime the pause state changes. Duplicated interval is prevented by implementing a cleanup function at the end of the hook.
I have tested it several times and it seems to be working normally. Hopefully I don't introduce anymore bugs with this current implementation...