Adding playlist page and topbar
- Adding user context
Fetches the logged user and provide the data to its children

- Readjust datetime converter
Previously, function in utils return a formatted string that can directly be used in jsx. Now, util function only return the necessary values and the formatting is done in the component.

- Set fixed min width and min height for the app
This decision is based on the spotify web app

- Change sidebar buttons to NavLink
Better navigation and enables dynamic styling based on active page